# More efficient serialization