Tested with [benchmarks from jsoniter-scala](https://github.com/plokhotnyuk/jsoniter-scala/tree/master/jsoniter-scala-benchmark/shared/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/benchmark):
## Before
```
[info] Benchmark                              (size)   Mode  Cnt         Score         Error  Units
[info] ADTWriting.zioJson                        N/A  thrpt    5    641749.429 ±   88832.341  ops/s
[info] AnyValsWriting.zioJson                    N/A  thrpt    5   1031663.091 ±  145986.261  ops/s
[info] ArrayOfBigDecimalsWriting.zioJson         128  thrpt    5     30140.787 ±    2272.644  ops/s
[info] ArrayOfBooleansWriting.zioJson            128  thrpt    5    162974.091 ±    7179.442  ops/s
[info] ArrayOfBytesWriting.zioJson               128  thrpt    5    146065.652 ±    5870.836  ops/s
[info] ArrayOfCharsWriting.zioJson               128  thrpt    5    143890.142 ±   13784.246  ops/s
[info] ArrayOfDoublesWriting.zioJson             128  thrpt    5     21607.773 ±    2188.168  ops/s
[info] ArrayOfDurationsWriting.zioJson           128  thrpt    5     41444.486 ±    2340.390  ops/s
[info] ArrayOfEnumADTsWriting.zioJson            128  thrpt    5    156212.569 ±   10641.182  ops/s
[info] ArrayOfEnumsWriting.zioJson               128  thrpt    5    164588.286 ±   22627.431  ops/s
[info] ArrayOfFloatsWriting.zioJson              128  thrpt    5     27790.314 ±    1507.760  ops/s
[info] ArrayOfInstantsWriting.zioJson            128  thrpt    5     20328.500 ±    2572.701  ops/s
[info] ArrayOfIntsWriting.zioJson                128  thrpt    5    121953.829 ±   11540.071  ops/s
[info] ArrayOfLocalDateTimesWriting.zioJson      128  thrpt    5     16427.644 ±     472.592  ops/s
[info] ArrayOfLocalDatesWriting.zioJson          128  thrpt    5     39251.896 ±    1746.547  ops/s
[info] ArrayOfLocalTimesWriting.zioJson          128  thrpt    5     22663.124 ±    2113.489  ops/s
[info] ArrayOfLongsWriting.zioJson               128  thrpt    5    101057.751 ±   11766.818  ops/s
[info] ArrayOfMonthDaysWriting.zioJson           128  thrpt    5     63572.597 ±    9639.537  ops/s
[info] ArrayOfOffsetDateTimesWriting.zioJson     128  thrpt    5     13868.431 ±    1017.160  ops/s
[info] ArrayOfOffsetTimesWriting.zioJson         128  thrpt    5     19015.918 ±    2607.406  ops/s
[info] ArrayOfPeriodsWriting.zioJson             128  thrpt    5     48070.910 ±    1858.225  ops/s
[info] ArrayOfShortsWriting.zioJson              128  thrpt    5    135105.655 ±   18431.134  ops/s
[info] ArrayOfUUIDsWriting.zioJson               128  thrpt    5     76567.762 ±    1907.395  ops/s
[info] ArrayOfYearMonthsWriting.zioJson          128  thrpt    5     61848.245 ±   10354.246  ops/s
[info] ArrayOfYearsWriting.zioJson               128  thrpt    5     65877.071 ±    6068.364  ops/s
[info] ArrayOfZoneIdsWriting.zioJson             128  thrpt    5     77841.887 ±    9743.622  ops/s
[info] ArrayOfZoneOffsetsWriting.zioJson         128  thrpt    5     81202.306 ±    4202.381  ops/s
[info] ArrayOfZonedDateTimesWriting.zioJson      128  thrpt    5     10003.821 ±     509.567  ops/s
[info] GitHubActionsAPIWriting.zioJson           N/A  thrpt    5    147843.742 ±    9461.001  ops/s
[info] GoogleMapsAPIWriting.zioJson              N/A  thrpt    5      9585.303 ±     572.394  ops/s
[info] IntWriting.zioJson                        N/A  thrpt    5  13726066.699 ± 1011886.445  ops/s
[info] ListOfBooleansWriting.zioJson             128  thrpt    5    173801.731 ±   21102.226  ops/s
[info] MapOfIntsToBooleansWriting.zioJson        128  thrpt    5     94150.682 ±   13055.820  ops/s
[info] PrimitivesWriting.zioJson                 N/A  thrpt    5   1203535.575 ±   48032.497  ops/s
[info] SetOfIntsWriting.zioJson                  128  thrpt    5    133599.520 ±   18815.939  ops/s
[info] StringOfAsciiCharsWriting.zioJson         128  thrpt    5   2003530.978 ±  143320.040  ops/s
[info] StringOfNonAsciiCharsWriting.zioJson      128  thrpt    5    940226.340 ±   51051.933  ops/s
[info] VectorOfBooleansWriting.zioJson           128  thrpt    5    178050.561 ±    7745.431  ops/s
```
## After
```
[info] Benchmark                              (size)   Mode  Cnt         Score         Error  Units
[info] ADTWriting.zioJson                        N/A  thrpt    5    794063.816 ±   71355.851  ops/s
[info] AnyValsWriting.zioJson                    N/A  thrpt    5   1822539.801 ±  355303.150  ops/s
[info] ArrayOfBigDecimalsWriting.zioJson         128  thrpt    5     33176.333 ±    4974.099  ops/s
[info] ArrayOfBooleansWriting.zioJson            128  thrpt    5    364610.489 ±   52042.210  ops/s
[info] ArrayOfBytesWriting.zioJson               128  thrpt    5    283862.207 ±   31803.045  ops/s
[info] ArrayOfCharsWriting.zioJson               128  thrpt    5    334642.594 ±   32364.070  ops/s
[info] ArrayOfDoublesWriting.zioJson             128  thrpt    5     24550.339 ±     207.527  ops/s
[info] ArrayOfDurationsWriting.zioJson           128  thrpt    5     64909.196 ±    3721.206  ops/s
[info] ArrayOfEnumADTsWriting.zioJson            128  thrpt    5    284354.820 ±    7252.089  ops/s
[info] ArrayOfEnumsWriting.zioJson               128  thrpt    5    319465.727 ±   47603.770  ops/s
[info] ArrayOfFloatsWriting.zioJson              128  thrpt    5     31281.729 ±    2498.736  ops/s
[info] ArrayOfInstantsWriting.zioJson            128  thrpt    5     52551.877 ±    8490.347  ops/s
[info] ArrayOfIntsWriting.zioJson                128  thrpt    5    231111.556 ±   38445.103  ops/s
[info] ArrayOfLocalDateTimesWriting.zioJson      128  thrpt    5     59553.638 ±    5692.185  ops/s
[info] ArrayOfLocalDatesWriting.zioJson          128  thrpt    5    115962.453 ±   11029.803  ops/s
[info] ArrayOfLocalTimesWriting.zioJson          128  thrpt    5     77035.734 ±    9184.439  ops/s
[info] ArrayOfLongsWriting.zioJson               128  thrpt    5    167704.343 ±   12354.562  ops/s
[info] ArrayOfMonthDaysWriting.zioJson           128  thrpt    5    145364.784 ±   17393.121  ops/s
[info] ArrayOfOffsetDateTimesWriting.zioJson     128  thrpt    5     47718.824 ±    3724.216  ops/s
[info] ArrayOfOffsetTimesWriting.zioJson         128  thrpt    5     60411.754 ±   12193.580  ops/s
[info] ArrayOfPeriodsWriting.zioJson             128  thrpt    5     81185.611 ±    9573.331  ops/s
[info] ArrayOfShortsWriting.zioJson              128  thrpt    5    270601.219 ±   25152.439  ops/s
[info] ArrayOfUUIDsWriting.zioJson               128  thrpt    5    135243.597 ±    5791.667  ops/s
[info] ArrayOfYearMonthsWriting.zioJson          128  thrpt    5    123175.013 ±   13964.292  ops/s
[info] ArrayOfYearsWriting.zioJson               128  thrpt    5    209178.366 ±   15524.710  ops/s
[info] ArrayOfZoneIdsWriting.zioJson             128  thrpt    5    256878.985 ±    8145.431  ops/s
[info] ArrayOfZoneOffsetsWriting.zioJson         128  thrpt    5    134306.787 ±   26935.020  ops/s
[info] ArrayOfZonedDateTimesWriting.zioJson      128  thrpt    5     41747.556 ±    4105.078  ops/s
[info] GitHubActionsAPIWriting.zioJson           N/A  thrpt    5    215886.849 ±   33569.250  ops/s
[info] GoogleMapsAPIWriting.zioJson              N/A  thrpt    5     17599.913 ±    1956.074  ops/s
[info] IntWriting.zioJson                        N/A  thrpt    5  19288438.667 ± 1768454.679  ops/s
[info] ListOfBooleansWriting.zioJson             128  thrpt    5    380134.457 ±    6992.076  ops/s
[info] MapOfIntsToBooleansWriting.zioJson        128  thrpt    5    104329.690 ±    9455.023  ops/s
[info] PrimitivesWriting.zioJson                 N/A  thrpt    5   2207577.744 ±  106832.946  ops/s
[info] SetOfIntsWriting.zioJson                  128  thrpt    5    238557.981 ±   46868.697  ops/s
[info] StringOfAsciiCharsWriting.zioJson         128  thrpt    5   1794317.936 ±  261846.541  ops/s
[info] StringOfNonAsciiCharsWriting.zioJson      128  thrpt    5    801937.772 ±   59733.380  ops/s
[info] VectorOfBooleansWriting.zioJson           128  thrpt    5    393926.863 ±   31341.238  ops/s
```